changed output from float to int

+++ onerepmax.py
@@ -15,5 +15,5 @@ brzycki=weight_lifted*(36/(37-exercise_reps))
 lombardi=weight_lifted*exercise_reps**0.1
 wendler=weight_lifted*exercise_reps*0.0333+weight_lifted
 
-one_rep_max=(brzycki+lombardi+wendler)/3
+one_rep_max=int((brzycki+lombardi+wendler)/3)
 print("Your one rep max estimate is: "+str(one_rep_max)+" pounds")
